+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Add warning modal [#2302](https://github.com/kvhnuke/etherwallet/pull/2302)
 - Sync tokens with ethereum-list [#2301](https://github.com/kvhnuke/etherwallet/pull/2301)
 
 ### Release v3.39.0